+++ crypto/protocol/src/protocol_transport/mod.rs
@@ -7,13 +7,17 @@ mod subscribe_message;
 use async_trait::async_trait;
 pub use broadcaster::Broadcaster;
 use errors::WsError;
-#[cfg(feature = "server")]
+#[cfg(any(feature = "server", feature = "wasm"))]
 use futures::{SinkExt, StreamExt};
 use noise::EncryptedWsConnection;
 pub use subscribe_message::SubscribeMessage;
+#[cfg(feature = "server")]
+use tokio::net::TcpStream;
 use tokio::sync::{broadcast, mpsc};
+#[cfg(feature = "server")]
+use tokio_tungstenite::{connect_async, tungstenite, MaybeTlsStream, WebSocketStream};
 
-use crate::{PartyId, ProtocolMessage};
+use crate::{errors::UserRunningProtocolErr, PartyId, ProtocolMessage};
 
 /// Channels between a remote party and the signing or DKG protocol
 pub struct WsChannels {
@@ -26,10 +30,34 @@ pub struct WsChannels {
 
 /// Represents the functionality of a Websocket connection with binary messages
 /// allowing us to generalize over different websocket implementations
-#[async_trait]
+#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 pub trait WsConnection {
     async fn recv(&mut self) -> Result<Vec<u8>, WsError>;
     async fn send(&mut self, msg: Vec<u8>) -> Result<(), WsError>;
+}
+
+#[cfg(feature = "wasm")]
+#[async_trait(?Send)]
+impl WsConnection for gloo_net::websocket::futures::WebSocket {
+    async fn recv(&mut self) -> Result<Vec<u8>, WsError> {
+        if let gloo_net::websocket::Message::Bytes(msg) = self
+            .next()
+            .await
+            .ok_or(WsError::ConnectionClosed)?
+            .map_err(|e| WsError::ConnectionError(e.to_string()))?
+        {
+            Ok(msg)
+        } else {
+            Err(WsError::UnexpectedMessageType)
+        }
+    }
+
+    async fn send(&mut self, msg: Vec<u8>) -> Result<(), WsError> {
+        SinkExt::send(&mut self, gloo_net::websocket::Message::Bytes(msg))
+            .await
+            .map_err(|_| WsError::ConnectionClosed)
+    }
 }
 
 #[cfg(feature = "server")]
@@ -54,9 +82,6 @@ impl WsConnection for axum::extract::ws::WebSocket {
             .map_err(|_| WsError::ConnectionClosed)
     }
 }
-
-#[cfg(feature = "server")]
-use tokio_tungstenite::{tungstenite, MaybeTlsStream};
 
 #[cfg(feature = "server")]
 #[async_trait]
@@ -96,18 +121,61 @@ pub async fn ws_to_channels<T: WsConnection>(
                 ws_channels.tx.send(msg).await.map_err(|_| WsError::MessageAfterProtocolFinish)?;
             }
             // Outgoing message (from signing protocol to remote peer)
-            Ok(msg) = ws_channels.broadcast.recv() => {
-                // Check that the message is for this peer
-                if let Some(party_id) = &msg.to {
-                    if party_id != &remote_party_id {
-                        continue;
+            msg_result = ws_channels.broadcast.recv() => {
+                if let Ok(msg) = msg_result {
+                    // Check that the message is for this peer
+                    if let Some(party_id) = &msg.to {
+                        if party_id != &remote_party_id {
+                            continue;
+                        }
                     }
+                    let message_vec = bincode::serialize(&msg)?;
+                    // TODO if this fails, the ws connection has been dropped during the protocol
+                    // we should inform the chain of this.
+                    connection.send(message_vec).await.map_err(|e| WsError::EncryptedConnection(e.to_string()))?;
+                } else {
+                    return Ok(());
                 }
-                let message_vec = bincode::serialize(&msg)?;
-                // TODO if this fails, the ws connection has been dropped during the protocol
-                // we should inform the chain of this.
-                connection.send(message_vec).await.map_err(|e| WsError::EncryptedConnection(e.to_string()))?;
             }
         }
     }
 }
+
+/// Open a ws connection in a native environment
+#[cfg(feature = "server")]
+pub async fn open_ws_connection(
+    address: String,
+) -> Result<WebSocketStream<MaybeTlsStream<TcpStream>>, UserRunningProtocolErr> {
+    let (ws_stream, _response) = connect_async(address)
+        .await
+        .map_err(|e| UserRunningProtocolErr::Connection(e.to_string()))?;
+    Ok(ws_stream)
+}
+
+/// Open a ws connection on wasm with the JS websocket API
+#[cfg(feature = "wasm")]
+pub async fn open_ws_connection(
+    address: String,
+) -> Result<gloo_net::websocket::futures::WebSocket, UserRunningProtocolErr> {
+    let ws_stream = gloo_net::websocket::futures::WebSocket::open(&address)
+        .map_err(|e| UserRunningProtocolErr::Connection(e.to_string()))?;
+    Ok(ws_stream)
+}
+
+// This dummy trait is only needed because we cant add #[cfg] to where clauses
+/// Trait only when not using wasm, adding the send marker trait
+#[cfg(feature = "server")]
+pub trait ThreadSafeWsConnection: WsConnection + std::marker::Send + 'static {}
+
+/// Trait only when using wasm, not adding the send marker trait
+#[cfg(feature = "wasm")]
+pub trait ThreadSafeWsConnection: WsConnection + 'static {}
+
+#[cfg(feature = "server")]
+impl ThreadSafeWsConnection
+    for WebSocketStream<tokio_tungstenite::MaybeTlsStream<tokio::net::TcpStream>>
+{
+}
+
+#[cfg(feature = "wasm")]
+impl ThreadSafeWsConnection for gloo_net::websocket::futures::WebSocket {}