+++ crypto/server/src/health/api.rs
@@ -1,5 +1,7 @@
 use axum::http::StatusCode;
 
+/// For checking the health of the TSS
+#[tracing::instrument]
 pub async fn healthz() -> StatusCode {
     StatusCode::OK
 }