+++ crypto/server/src/user/tests.rs
@@ -19,9 +19,14 @@ use futures::{
     join, Future, SinkExt, StreamExt,
 };
 use hex_literal::hex;
-use kvdb::{clean_tests, encrypted_sled::PasswordMethod, kv_manager::value::KvManager};
+use kvdb::{
+    clean_tests,
+    encrypted_sled::PasswordMethod,
+    kv_manager::{helpers::deserialize as keyshare_deserialize, value::KvManager},
+};
 use more_asserts as ma;
 use parity_scale_codec::Encode;
+use serde::{Deserialize, Serialize};
 use serial_test::serial;
 use sp_core::{crypto::Ss58Codec, Pair as OtherPair, H160};
 use sp_keyring::{AccountKeyring, Sr25519Keyring};
@@ -50,7 +55,10 @@ use testing_utils::{
     },
     test_client,
 };
-use tokio::task::JoinHandle;
+use tokio::{
+    io::{AsyncRead, AsyncReadExt},
+    task::JoinHandle,
+};
 use tokio_tungstenite::{connect_async, tungstenite::Message};
 use x25519_dalek::{PublicKey, StaticSecret};
 
@@ -60,18 +68,19 @@ use crate::{
     get_signer,
     helpers::{
         launch::{
-            setup_mnemonic, Configuration, ValidatorName, DEFAULT_BOB_MNEMONIC,
+            load_kv_store, setup_mnemonic, Configuration, ValidatorName, DEFAULT_BOB_MNEMONIC,
             DEFAULT_CHARLIE_MNEMONIC, DEFAULT_ENDPOINT, DEFAULT_MNEMONIC,
         },
         signing::{create_unique_tx_id, Hasher},
         substrate::{get_subgroup, make_register, return_all_addresses_of_subgroup},
         tests::{
-            check_if_confirmation, create_clients, keyring_to_subxt_signer_and_x25519,
-            run_to_block, setup_client, spawn_testing_validators, update_programs,
+            check_if_confirmation, create_clients, initialize_test_logger,
+            keyring_to_subxt_signer_and_x25519, run_to_block, setup_client,
+            spawn_testing_validators, update_programs,
         },
         user::send_key,
     },
-    load_kv_store, new_user,
+    new_user,
     r#unsafe::api::UnsafeQuery,
     signing_client::ListenerState,
     user::api::{recover_key, UserRegistrationInfo, UserSignatureRequest},
@@ -82,7 +91,9 @@ use crate::{
 #[tokio::test]
 #[serial]
 async fn test_get_signer_does_not_throw_err() {
+    initialize_test_logger();
     clean_tests();
+
     let kv_store = load_kv_store(&None, false).await;
     let mnemonic = setup_mnemonic(&kv_store, &None).await;
     assert!(mnemonic.is_ok());
@@ -92,7 +103,9 @@ async fn test_get_signer_does_not_throw_err() {
 #[tokio::test]
 #[serial]
 async fn test_sign_tx_no_chain() {
+    initialize_test_logger();
     clean_tests();
+
     let one = AccountKeyring::Dave;
     let two = AccountKeyring::Two;
 
@@ -350,7 +363,9 @@ async fn test_sign_tx_no_chain() {
 #[tokio::test]
 #[serial]
 async fn test_fail_signing_group() {
+    initialize_test_logger();
     clean_tests();
+
     let dave = AccountKeyring::Dave;
     let _ = spawn_testing_validators(None, false).await;
 
@@ -403,7 +418,9 @@ async fn test_fail_signing_group() {
 #[tokio::test]
 #[serial]
 async fn test_store_share() {
+    initialize_test_logger();
     clean_tests();
+
     let alice = AccountKeyring::Alice;
     let alice_program = AccountKeyring::Charlie;
     let signing_address = alice.to_account_id().to_ss58check();
@@ -556,6 +573,8 @@ async fn test_store_share() {
 #[tokio::test]
 #[serial]
 async fn test_return_addresses_of_subgroup() {
+    initialize_test_logger();
+
     let cxt = test_context_stationary().await;
     let api = get_api(&cxt.node_proc.ws_url).await.unwrap();
     let rpc = get_rpc(&cxt.node_proc.ws_url).await.unwrap();
@@ -567,7 +586,9 @@ async fn test_return_addresses_of_subgroup() {
 #[tokio::test]
 #[serial]
 async fn test_send_and_receive_keys() {
+    initialize_test_logger();
     clean_tests();
+
     let alice = AccountKeyring::Alice;
 
     let cxt = test_context_stationary().await;
@@ -660,7 +681,9 @@ async fn test_send_and_receive_keys() {
 #[tokio::test]
 #[serial]
 async fn test_recover_key() {
+    initialize_test_logger();
     clean_tests();
+
     let cxt = test_node_process_testing_state(false).await;
     setup_client().await;
     let (_, bob_kv) =
@@ -718,7 +741,9 @@ pub async fn put_register_request_on_chain(
 #[tokio::test]
 #[serial]
 async fn test_sign_tx_user_participates() {
+    initialize_test_logger();
     clean_tests();
+
     let one = AccountKeyring::Eve;
     let two = AccountKeyring::Two;
 
@@ -1019,6 +1044,7 @@ async fn test_sign_tx_user_participates() {
 #[tokio::test]
 #[serial]
 async fn test_register_with_private_key_visibility() {
+    initialize_test_logger();
     clean_tests();
 
     let one = AccountKeyring::One;
@@ -1085,6 +1111,7 @@ async fn test_register_with_private_key_visibility() {
     assert_eq!(response.text().await.unwrap(), "");
 
     assert!(keyshare_result.is_ok());
+    clean_tests();
 }
 
 // This tests the test client included in testing_utils.