+++ crypto/server/src/version/api.rs
@@ -1,0 +1,5 @@
+/// Returns the version and commit data
+#[tracing::instrument]
+pub async fn version() -> String {
+    format!("{}-{}", env!("VERGEN_RUSTC_SEMVER"), env!("VERGEN_GIT_DESCRIBE"))
+}