+++ crypto/shared/src/lib.rs
@@ -18,3 +18,6 @@ pub const PRUNE_BLOCK: u32 = 14400;
 
 /// Timeout for validators to wait for other validators to join protocol committees
 pub const SETUP_TIMEOUT_SECONDS: u64 = 20;
+
+/// The amount of proactive refreshes we do per session
+pub const REFRESHES_PER_SESSION: u32 = 10;